pyproject.toml
[project]
name = "mlops-vehicle-insurance-claims"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "catboost>=1.2.10",
    "click>=8.3.2",
    "memray>=1.15.0",
    "mlxtend>=0.23.4",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pyinstrument>=5.0.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "setuptools>=70,<82",  # pkg_resources (ydata-profiling) removed in setuptools 82+
    "tqdm>=4.67.3",
    "ydata-profiling>=4.18.3",
]
